Move some common functionality.
