--- hlavo/ingress/moist_profile/requirements.txt
@@ -1,10 +1,0 @@
-matplotlib
-pandas
-
-numpy
-scipy
-
-# web DataFlow grabbing
-python-dotenv
-selenium
-requests
